ch11/myproject_virtualenv/src/django-myproject/requirements/_base.txt
appnope==0.1.0
asgiref==3.2.3
backcall==0.1.0
blessings==1.7
bpython==0.18
certifi==2019.11.28
chardet==3.0.4
coverage==4.5.3
curtsies==0.3.0
decorator==4.4.1
Django==3.0.1
django-debug-toolbar==2.1
greenlet==0.4.15
idna==2.8
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2019.2
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
transliterate==1.10.2
typing==3.7.4.1
urllib3==1.25.7
wcwidth==0.1.7
